Move manageConfigs helper into separate file
